requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.7.0
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bidict==0.22.1
billiard==4.1.0
blinker==1.6.2
blis==0.7.9
boilerpy3==1.0.6
CacheControl==0.13.1
cachetools==5.3.1
canals==0.2.2
catalogue==2.0.8
cattrs==23.1.2
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorlog==6.7.0
comm==0.1.3
confection==0.1.0
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.9
datasets==2.13.1
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.6
dnspython==2.3.0
docopt==0.6.2
elastic-transport==8.4.0
elasticsearch==8.8.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl#sha256=c8ac64840c1eb3e3ca7bd38bd1e1c48fb0faeb2449d54d01d5ce629af4595775
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
evaluate==0.4.0
eventlet==0.33.3
Events==0.4
exceptiongroup==1.1.2
executing==1.2.0
faiss-cpu==1.7.2
farm-haystack==1.18.1
filelock==3.12.2
firebase-admin==6.2.0
Flask==1.1.2
Flask-Executor==1.0.0
Flask-ReCaptcha==0.4.2
Flask-SocketIO==5.3.5
flatbuffers==23.5.26
fonttools==4.41.1
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.4.0
gcloud==0.18.3
google-api-core==1.34.0
google-api-python-client==2.70.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
googletrans==4.0.0rc1
gpt4all==1.0.1
greenlet==2.0.2
grpcio==1.56.0
grpcio-status==1.48.2
grpcio-tools==1.56.0
h11==0.9.0
h2==3.2.0
h5py==3.9.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.15.1
hyperframe==5.2.0
idna==2.10
inflect==7.0.0
ipykernel==6.24.0
ipython==8.14.0
itsdangerous==2.0.1
jaraco.context==4.3.0
jedi==0.18.2
Jinja2==2.11.2
joblib==1.3.1
jq==1.4.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jwcrypto==1.5.0
jws==0.1.3
keras==2.13.1
kiwisolver==1.4.4
kombu==5.3.1
langchain==0.0.226
langchainplus-sdk==0.0.20
langcodes==3.3.0
lazy-imports==0.3.1
libclang==16.0.0
loguru==0.7.0
manifest-ml==0.0.1
Markdown==3.4.3
MarkupSafe==2.0.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
num2words==0.5.12
numexpr==2.8.4
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
openapi-schema-pydantic==1.2.4
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pinecone-client==2.2.2
platformdirs==3.8.0
posthog==3.0.1
preshed==3.0.8
prompt-toolkit==3.0.38
prompthub-py==4.0.0
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.10
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
Pyrebase4==4.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-jwt==2.0.1
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
qdrant-client==0.11.10
quantulum3==0.9.0
rank-bm25==0.2.2
redis==4.6.0
referencing==0.29.1
regex==2023.6.3
requests==2.29.0
requests-cache==0.9.8
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
responses==0.18.0
rfc3986==1.5.0
rpds-py==0.8.10
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setfit==0.7.0
simplejson==3.19.1
six==1.16.0
sklearn==0.0.post5
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.48
SQLAlchemy-Utils==0.41.1
sqlitedict==2.1.0
srsly==2.4.6
sseclient-py==1.7.2
stack-data==0.6.2
sympy==1.12
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
tensorflow-metal==1.0.1
termcolor==2.3.0
textblob==0.17.1
thinc==8.1.10
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.25.11
validators==0.20.0
vine==5.0.0
wasabi==1.1.2
wcwidth==0.2.6
weaviate-client==3.21.0
Werkzeug==2.0.3
wikipedia==1.4.0
wolframalpha==5.0.0
wrapt==1.15.0
WTForms==3.0.1
xgboost==1.7.6
xmltodict==0.13.0
xxhash==3.2.0
yarl==1.9.2
